corruption rounding improved

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/wands/WandOfRegrowth.java
@@ -119,7 +119,7 @@ public class WandOfRegrowth extends Wand {
 						Statistics.qualifiedForBossChallengeBadge = false;
 					}
 					
-					float duration = 4f * chrgUsed;
+					int duration = 4 * chrgUsed;
 					if(ch.buff(ChampionEnemy.AntiMagic.class) != null) {
 						duration /= 2;
 					}
